@@ -1,5 +1,3 @@
 module github.com/rbranson/camo
 
-go 1.17
-
-require github.com/google/go-cmp v0.5.7
+go 1.21
